@@ -1,6 +1,6 @@
 package com.frogobox.kamusapps.views.activities;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.widget.TextView;
 
