@@ -1,7 +1,7 @@
 package com.frogobox.kamusapps.views.activities;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.ProgressBar;
 
 import com.frogobox.kamusapps.R;
